aural/pyproject.toml
[project]
name = "aural"
version = "0.1.0"
description = "Audio content extraction and Stashapp import system for GoneWildAudio content"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31.0",
    "praw>=7.7.1",
    "python-dotenv>=1.0.0",
    "stashapp-tools>=0.2.59",
    "playwright>=1.57.0",
    "httpx>=0.28.0",
    "cryptography>=44.0.0",
    "beautifulsoup4>=4.12.0",
    "yt-dlp>=2025.12.8",
    "oshash>=0.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include all Python files in the root (this is a scripts collection, not a package)
include = ["*.py"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
